% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calcul_distribution_age.R
\name{calcul_distribution_age}
\alias{calcul_distribution_age}
\title{Calcul_distribution_age}
\usage{
calcul_distribution_age(data)
}
\arguments{
\item{data}{Un data.frame contenant une colonne "Date.de.naissance" avec les dates de naissance
des individus au format "jour/mois/année".}
}
\value{
Un vecteur numérique contenant les quantiles des âges des individus dans le jeu de données.
Les quantiles sont calculés pour les probabilités suivantes : 0, 0.25, 0.5, 0.75, et 1.
}
\description{
Cette fonction calcule la distribution des âges des individus dans un jeu de données
en utilisant la date de naissance. Elle retourne les quantiles des âges des individus,
allant du minimum au maximum, ainsi que les quantiles à 25\%, 50\% (médiane) et 75\%.
}
