% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compter_nombre_adjoints.R
\name{compter_nombre_d_adjoints}
\alias{compter_nombre_d_adjoints}
\title{Compte_Nombre_d_adjoints}
\usage{
compter_nombre_d_adjoints(data)
}
\arguments{
\item{data}{un dataframe qui contient les informations des élus.}
}
\value{
un entier correspondant au nombre d'adjoints.
}
\description{
Cette fonction compte le nombre d'adjoints
}
