% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_commune.R
\name{creer_commune}
\alias{creer_commune}
\title{Creer_commune}
\usage{
creer_commune(data, commune_name)
}
\arguments{
\item{data}{Un \code{data.frame} contenant des informations sur les communes.
Il doit inclure la colonne 'Libellé.de.la.commune'.}

\item{commune_name}{Nom de la commune que vous voulez extraire du \code{data.frame}. Rappellez-vous de ne pas ajouter le numéro de la commune.}
}
\value{
Un \code{data.frame} de la classe 'commune' créé à partir du \code{data.frame} d'entrée.
}
\description{
Cette fonction prend un \code{data.frame} en entrée et renvoie un \code{data.frame}
avec la classe 'commune'. Le \code{data.frame} d'entrée doit inclure la colonne
'Libellé.de.la.commune' et cette fonction va extraire les lignes correspondant
à une commune spécifique.
}
\examples{
# Exemple d'utilisation de la fonction creer_commune
df <- data.frame(
  Code.du.département = c("01", "01", "02", "02"),
  Libellé.du.département = c("Ain", "Ain", "Aisne", "Aisne"),
  Libellé.de.la.commune = c("Lyon", "Lyon", "Soissons", "Soissons"),
  Nom.de.l.élu = c("Dupont", "Martin", "Lemoine", "Dubois")
)
commune_obj <- creer_commune(df, "Lyon")  # Sélectionne les données pour la commune "Lyon"
print(commune_obj)  # Affiche le data.frame de la commune avec la classe 'commune'
}
