% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_departement.R
\name{creer_departement}
\alias{creer_departement}
\title{Creer_departement}
\usage{
creer_departement(data, departement_name)
}
\arguments{
\item{data}{Un \code{data.frame} contenant des informations sur les départements.
Il doit inclure la colonne 'Libellé.du.département'.}

\item{departement_name}{Nom du département que vous voulez extraire du \code{data.frame}. Rappellez-vous de ne pas ajouter le numéro du département.}
}
\value{
Un \code{data.frame} de la classe 'departement' créé à partir du \code{data.frame} d'entrée.
}
\description{
Cette fonction prend un \code{data.frame} en entrée et renvoie un \code{data.frame}
avec la classe 'departement'. Le \code{data.frame} d'entrée doit inclure la colonne
'Libellé.du.département' et cette fonction va extraire les lignes correspondant
à un département spécifique.
}
\examples{
# Exemple d'utilisation de la fonction creer_departement
df <- data.frame(
  Code.du.département = c("01", "01", "02", "02"),
  Libellé.du.département = c("Ain", "Ain", "Aisne", "Aisne"),
  Libellé.de.la.commune = c("Lyon", "Lyon", "Soissons", "Soissons"),
  Nom.de.l.élu = c("Dupont", "Martin", "Lemoine", "Dubois")
)
departement_obj <- creer_departement(df, "Ain")  # Sélectionne les données pour le département "Ain"
print(departement_obj)  # Affiche le data.frame du département avec la classe 'departement'
}
