% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elus_cm.R
\docType{data}
\name{elus_cm}
\alias{elus_cm}
\title{Dataset des élus}
\format{
\subsection{'elus_cm'}{

Un dataframe composé de 14 colonnes et 489972 lignes.
\describe{
\item{Code.du.département}{Code du département où l'élu.e exerce.}
\item{Libellé.du.département}{Nom du département où l'élu.e exerce.}
\item{Code.de.la.commune}{Code de la commune où l'élu.e exerce.}
\item{Libellé.de.la.commune}{Nom de la commune où l'élu.e exerce.}
\item{Nom.de.l.élu}{Nom de l'élu.e.}
\item{Prénom.de.l.élu}{Prénom de l'élu.e.}
\item{Code.sexe}{M pour les hommes et F pour les femmes.}
\item{Date.de.naissance}{Date de naissance de l'élu.e au format "jour/mois/année".}
\item{Code.de.la.catégorie.socio.professionnelle}{Code indiquant la catégorie socio-professionnelle de l'élu.e.}
\item{Libellé.de.la.catégorie.socio.professionnelle}{Libellé de la catégorie socio-professionnelle de l'élu.e.}
\item{Date.de.début.du.mandat}{Date de début du mandat de l'élu.e.}
\item{Libellé.de.la.fonction}{Fonction de l'élu.e (par exemple, maire, conseiller).}
\item{Date.de.début.de.la.fonction}{Date de début de la fonction de l'élu.e.}
\item{Code.nationalité}{Code de la nationalité de l'élu.e.}
}
}
}
\source{
\url{https://www.data.gouv.fr}
}
\usage{
elus_cm
}
\description{
Ce dataset contient des informations sur les élus d'une ou plusieurs communes et departéments en France, y compris leur nom, prénom, date de naissance,
catégorie socio-professionnelle, et d'autres informations concernant leur mandat.
}
\keyword{datasets}
