% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_professions.R
\name{plot_code_professions}
\alias{plot_code_professions}
\title{Plot_professions}
\usage{
plot_code_professions(data)
}
\arguments{
\item{data}{Un data.frame contenant les informations suivantes :
\itemize{
\item "Libellé.de.la.catégorie.socio.professionnelle" : La catégorie socio-professionnelle de l'élu.
}}
}
\value{
Un graphique en barres horizontal affichant la distribution des codes socio-professionnels.
}
\description{
Cette fonction génère un graphique en barres horizontal représentant la distribution
des codes socio-professionnels des élus dans un data.frame. Elle utilise la colonne
"Libellé.de.la.catégorie.socio.professionnelle" pour créer un tableau de fréquences et
affiche un barplot de ces fréquences.
}
\examples{
# Exemple d'utilisation
data <- data.frame(
  Libellé.de.la.catégorie.socio.professionnelle = c("Enseignant", "Médecin", "Ingénieur", "Médecin", "Enseignant")
)
plot_code_professions(data)

}
