% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_commune.R
\name{plot_commune}
\alias{plot_commune}
\title{Plot_commune}
\usage{
plot_commune(data)
}
\arguments{
\item{data}{Un data.frame contenant les informations des élu.e.s. Il doit inclure les colonnes suivantes :
\itemize{
\item "Libellé.de.la.commune" : Le nom de la commune.
\item "Libellé.du.département" : Le nom du département.
\item "Libellé.de.la.catégorie.socio.professionnelle" : La catégorie socio-professionnelle des élu.e.s.
}}
}
\value{
Aucun retour, mais affiche un graphique en barres horizontal.
}
\description{
Cette fonction génère un graphique en barres pour visualiser la distribution des catégories socio-professionnelles des élu.e.s dans une commune spécifique.
Le titre du graphique sera le nom de la commune suivi du nom de son département.
L'axe des abscisses s'appelle "Libellés des codes professionnels pour les élus", où X est remplacé par le nombre d'élus de la commune.
}
\examples{
# Exemple d'utilisation ajusté à la fonction plot_commune
# Créons un data.frame de données fictives
df <- data.frame(
  Code.du.département = c("01", "01", "01", "01", "02", "02"),
  Libellé.du.département = c("Ain", "Ain", "Ain", "Ain", "Aisne", "Aisne"),
  Libellé.de.la.commune = c("Lyon", "Lyon", "Lyon", "Lyon", "Soissons", "Soissons"),
  Nom.de.l.élu = c("Dupont", "Martin", "Lemoine", "Bernard", "Dubois", "Lemoine"),
  Prénom.de.l.élu = c("Jean", "Pierre", "Luc", "Sophie", "Marc", "Claire"),
  Libellé.de.la.catégorie.socio.professionnelle = c("Médecin", "Ingénieur", "Enseignant", "Médecin", "Médecin", "Ingénieur")
)

# Filtrons le data.frame pour ne contenir que les données d'une commune spécifique
df_lyon <- df[df$Libellé.de.la.commune == "Lyon", ]

# Appel de la fonction plot_commune avec le data.frame df_lyon
plot_commune(df_lyon)

}
