% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_departement.R
\name{plot_departement}
\alias{plot_departement}
\title{plot_departement}
\usage{
plot_departement(data)
}
\arguments{
\item{data}{Un data.frame contenant les informations des élu.e.s. Il doit inclure les colonnes suivantes :
\itemize{
\item "Libellé.du.département" : Le nom du département.
\item "Libellé.de.la.catégorie.socio.professionnelle" : La catégorie socio-professionnelle des élu.e.s.
\item "Libellé.de.la.commune" : Le nom de la commune.
}}
}
\value{
Aucun retour, mais affiche un graphique en barres horizontal pour chaque département dans les données.
}
\description{
Cette fonction génère des graphiques en barres horizontales pour visualiser les 10 catégories socio-professionnelles les plus représentées dans un département donné.
Le titre du graphique sera le nom du département, suivi du nombre de communes dans ce département.
L'axe des abscisses s'appelle : "Libellés des 10 codes professionnels les plus représentés pour le département", où X est remplacé par le nom du département.
}
\examples{
# Exemple d'utilisation ajusté à la fonction plot_departement
# Créons un data.frame de données fictives
df <- data.frame(
  Code.du.département = c("01", "01", "01", "01", "02", "02"),
  Libellé.du.département = c("Ain", "Ain", "Ain", "Ain", "Aisne", "Aisne"),
  Libellé.de.la.commune = c("Lyon", "Lyon", "Lyon", "Lyon", "Soissons", "Soissons"),
  Nom.de.l.élu = c("Dupont", "Martin", "Lemoine", "Bernard", "Dubois", "Lemoine"),
  Libellé.de.la.catégorie.socio.professionnelle = c("Médecin", "Ingénieur", "Enseignant", "Médecin", "Médecin", "Ingénieur")
)

# Filtrons le data.frame pour qu'il ne contienne que les données d'un département spécifique
df_ain <- df[df$Libellé.du.département == "Ain", ]

# Appel de la fonction plot_departement avec le data.frame df_ain
plot_departement(df_ain)

}
