% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary_department.R
\name{Summary_departement}
\alias{Summary_departement}
\title{Summary_department}
\usage{
Summary_departement(x)
}
\arguments{
\item{x}{Un data.frame contenant les informations des élus. Il doit avoir les colonnes suivantes :
\itemize{
\item "Code.du.département" : Le code du département.
\item "Libellé.du.département" : Le nom du département.
\item "Libellé.de.la.commune" : Le nom de la commune.
\item "Nom.de.l.élu" : Le nom de l'élu.
\item "Prénom.de.l.élu" : Le prénom de l'élu.
\item "Date.de.naissance" : La date de naissance de l'élu (au format "jour/mois/année").
}}
}
\value{
Aucune valeur n'est retournée, mais la fonction affiche les résultats suivants :
\itemize{
\item Le nom du département.
\item Le nombre de communes dans le département.
\item Le nombre d'élu.e.s dans le département.
\item La distribution des âges des élu.e.s.
\item Le nom, le prénom, l'âge et la commune de l'élu.e le plus âgé(e).
\item Le nom, le prénom, l'âge et la commune de l'élu.e le plus jeune.
\item Le nom de la commune ayant la moyenne d'âge la plus faible, ainsi que la distribution des âges des élu.e.s de cette commune.
\item Le nom de la commune ayant la moyenne d'âge la plus élevée, ainsi que la distribution des âges des élu.e.s de cette commune.
}
}
\description{
Cette fonction génère un résumé statistique des élus d'un département, y compris les informations
sur les âges, les élus les plus âgés et les plus jeunes, ainsi que les communes ayant les moyennes
d'âge les plus faibles et les plus élevées.
}
\examples{
# Exemple d'utilisation
# Supposons que df est un dataframe contenant les données nécessaires
df <- data.frame(
  `Code.du.département` = c("01", "01", "01"),
  `Libellé.du.département` = c("Ain", "Ain", "Ain"),
  `Libellé.de.la.commune` = c("Bourg-en-Bresse", "Ambérieu-en-Bugey", "Lagnieu"),
  `Nom.de.l.élu` = c("Dupont", "Martin", "Bernard"),
  `Prénom.de.l.élu` = c("Pierre", "Jacques", "Lucie"),
  `Date.de.naissance` = c("01/01/1970", "15/03/1980", "23/06/1990")
)

# Appel de la fonction avec le dataframe exemple
Summary_departement(df)

}
