% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trouver_l_elu_le_plus_age.R
\name{trouver_l_elu_le_plus_age}
\alias{trouver_l_elu_le_plus_age}
\title{Trouver l'élu le plus âgé}
\usage{
trouver_l_elu_le_plus_age(data)
}
\arguments{
\item{data}{Un data.frame contenant les colonnes suivantes :
\itemize{
\item "Nom.de.l.élu" : Le nom de l'élu.
\item "Prénom.de.l.élu" : Le prénom de l'élu.
\item "Date.de.naissance" : La date de naissance de l'élu au format "jour/mois/année".
}}
}
\value{
Un data.frame contenant les informations de l'élu le plus âgé, incluant :
\itemize{
\item "Nom.de.l.élu" : Le nom de l'élu.
\item "Prénom.de.l.élu" : Le prénom de l'élu.
\item "Age" : L'âge de l'élu en années.
}
}
\description{
Cette fonction prend un data.frame contenant des informations sur les élus, y compris leur date de naissance,
et retourne les informations de l'élu le plus âgé.
}
\examples{
# Exemple d'utilisation
data <- data.frame(
  Nom.de.l.élu = c("Dupont", "Lemoine", "Dufresne"),
  Prénom.de.l.élu = c("Pierre", "Sophie", "Marc"),
  Date.de.naissance = c("01/01/1970", "15/05/1980", "23/07/1965")
)
# Appeler la fonction pour trouver l'élu le plus âgé
trouver_l_elu_le_plus_age(data)
# Retourne : Dupont, Pierre, 55

}
